requirements.txt
Flask
flask-bcrypt
tensorflow
opencv-python
numpy
pymongo
huggingface-hub
xhtml2pdf
cloudinary
python-dotenv
pytz
gunicorn